refactor MealRecord: remove subscriptions and model definitions
- Removed GraphQL subscriptions for MealRecord (create, update, delete) from subscriptions.ts.
- Deleted MealRecord type definitions from index.d.ts and index.js.
- Eliminated MealRecord schema definition from schema.js.
- Updated ui-components index.js to remove MealRecord forms.
