+++ src/graphql/mutations.ts
@@ -245,96 +245,6 @@ export const deleteDailyMealRecord = /* GraphQL */ `mutation DeleteDailyMealReco
   APITypes.DeleteDailyMealRecordMutationVariables,
   APITypes.DeleteDailyMealRecordMutation
 >;
-export const createMealRecord = /* GraphQL */ `mutation CreateMealRecord(
-  $input: CreateMealRecordInput!
-  $condition: ModelMealRecordConditionInput
-) {
-  createMealRecord(input: $input, condition: $condition) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedMutation<
-  APITypes.CreateMealRecordMutationVariables,
-  APITypes.CreateMealRecordMutation
->;
-export const updateMealRecord = /* GraphQL */ `mutation UpdateMealRecord(
-  $input: UpdateMealRecordInput!
-  $condition: ModelMealRecordConditionInput
-) {
-  updateMealRecord(input: $input, condition: $condition) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedMutation<
-  APITypes.UpdateMealRecordMutationVariables,
-  APITypes.UpdateMealRecordMutation
->;
-export const deleteMealRecord = /* GraphQL */ `mutation DeleteMealRecord(
-  $input: DeleteMealRecordInput!
-  $condition: ModelMealRecordConditionInput
-) {
-  deleteMealRecord(input: $input, condition: $condition) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedMutation<
-  APITypes.DeleteMealRecordMutationVariables,
-  APITypes.DeleteMealRecordMutation
->;
 export const createUserMealPreset = /* GraphQL */ `mutation CreateUserMealPreset(
   $input: CreateUserMealPresetInput!
   $condition: ModelUserMealPresetConditionInput
