+++ src/graphql/queries.ts
@@ -203,93 +203,6 @@ export const syncDailyMealRecords = /* GraphQL */ `query SyncDailyMealRecords(
   APITypes.SyncDailyMealRecordsQueryVariables,
   APITypes.SyncDailyMealRecordsQuery
 >;
-export const getMealRecord = /* GraphQL */ `query GetMealRecord($id: ID!) {
-  getMealRecord(id: $id) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedQuery<
-  APITypes.GetMealRecordQueryVariables,
-  APITypes.GetMealRecordQuery
->;
-export const listMealRecords = /* GraphQL */ `query ListMealRecords(
-  $filter: ModelMealRecordFilterInput
-  $limit: Int
-  $nextToken: String
-) {
-  listMealRecords(filter: $filter, limit: $limit, nextToken: $nextToken) {
-    items {
-      id
-      date
-      category
-      createdAt
-      updatedAt
-      _version
-      _deleted
-      _lastChangedAt
-      owner
-      __typename
-    }
-    nextToken
-    startedAt
-    __typename
-  }
-}
-` as GeneratedQuery<
-  APITypes.ListMealRecordsQueryVariables,
-  APITypes.ListMealRecordsQuery
->;
-export const syncMealRecords = /* GraphQL */ `query SyncMealRecords(
-  $filter: ModelMealRecordFilterInput
-  $limit: Int
-  $nextToken: String
-  $lastSync: AWSTimestamp
-) {
-  syncMealRecords(
-    filter: $filter
-    limit: $limit
-    nextToken: $nextToken
-    lastSync: $lastSync
-  ) {
-    items {
-      id
-      date
-      category
-      createdAt
-      updatedAt
-      _version
-      _deleted
-      _lastChangedAt
-      owner
-      __typename
-    }
-    nextToken
-    startedAt
-    __typename
-  }
-}
-` as GeneratedQuery<
-  APITypes.SyncMealRecordsQueryVariables,
-  APITypes.SyncMealRecordsQuery
->;
 export const getUserMealPreset = /* GraphQL */ `query GetUserMealPreset($id: ID!) {
   getUserMealPreset(id: $id) {
     id
