+++ src/graphql/subscriptions.ts
@@ -245,96 +245,6 @@ export const onDeleteDailyMealRecord = /* GraphQL */ `subscription OnDeleteDaily
   APITypes.OnDeleteDailyMealRecordSubscriptionVariables,
   APITypes.OnDeleteDailyMealRecordSubscription
 >;
-export const onCreateMealRecord = /* GraphQL */ `subscription OnCreateMealRecord(
-  $filter: ModelSubscriptionMealRecordFilterInput
-  $owner: String
-) {
-  onCreateMealRecord(filter: $filter, owner: $owner) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedSubscription<
-  APITypes.OnCreateMealRecordSubscriptionVariables,
-  APITypes.OnCreateMealRecordSubscription
->;
-export const onUpdateMealRecord = /* GraphQL */ `subscription OnUpdateMealRecord(
-  $filter: ModelSubscriptionMealRecordFilterInput
-  $owner: String
-) {
-  onUpdateMealRecord(filter: $filter, owner: $owner) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedSubscription<
-  APITypes.OnUpdateMealRecordSubscriptionVariables,
-  APITypes.OnUpdateMealRecordSubscription
->;
-export const onDeleteMealRecord = /* GraphQL */ `subscription OnDeleteMealRecord(
-  $filter: ModelSubscriptionMealRecordFilterInput
-  $owner: String
-) {
-  onDeleteMealRecord(filter: $filter, owner: $owner) {
-    id
-    date
-    category
-    foods {
-      id
-      name
-      calories
-      protein
-      carbohydrates
-      fat
-      __typename
-    }
-    createdAt
-    updatedAt
-    _version
-    _deleted
-    _lastChangedAt
-    owner
-    __typename
-  }
-}
-` as GeneratedSubscription<
-  APITypes.OnDeleteMealRecordSubscriptionVariables,
-  APITypes.OnDeleteMealRecordSubscription
->;
 export const onCreateUserMealPreset = /* GraphQL */ `subscription OnCreateUserMealPreset(
   $filter: ModelSubscriptionUserMealPresetFilterInput
   $owner: String
